+++ test/PerpetualIntelligence.Terminal.Test/Configuration/Options/TerminalOptionsTests.cs
@@ -24,9 +24,8 @@ namespace PerpetualIntelligence.Terminal.Configuration.Options
             TerminalOptions options = new();
 
             options.Id.Should().BeEmpty();
-            options.Name.Should().BeNull();
-            options.RootAsDriver.Should().BeFalse();
 
+            options.Driver.Should().NotBeNull();
             options.Authentication.Should().NotBeNull();
             options.Checker.Should().NotBeNull();
             options.Extractor.Should().NotBeNull();